refactor: PoC for .graphql imports
