+++ addon/graphql.d.ts
@@ -1,0 +1,6 @@
+declare module '*.graphql' {
+  import { DocumentNode } from 'graphql';
+  const defaultDocument: DocumentNode;
+
+  export default defaultDocument;
+}
