+++ tests/dummy/app/routes/characters.js
@@ -1,6 +1,6 @@
 import Route from '@ember/routing/route';
 import { RouteQueryManager } from 'ember-apollo-client';
-import query from 'dummy/gql/queries/characters';
+import query from 'dummy/gql/queries/characters.graphql';
 
 export default Route.extend(RouteQueryManager, {
   queryParams: {
