+++ tests/unit/services/apollo-test.ts
@@ -2,8 +2,8 @@ import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 import { computed } from '@ember-decorators/object';
 import ApolloService from 'ember-apollo-client/services/apollo';
-import testQuery from '../build/test-query';
-import testMutation from '../build/test-mutation';
+import testQuery from '../build/test-query.graphql';
+import testMutation from '../build/test-mutation.graphql';
 import { Promise } from 'rsvp';
 import getSuperComputed from 'dummy/tests/helpers/call-super-computed';
 import { InMemoryCache } from 'apollo-cache-inmemory';
